// resources/js/api/generated/models/getApiV1ServersServeridStacksStacknameFilesStats200.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1ServersServeridStacksStacknameFilesStats200 = {
  group_name?: string;
  /** @minimum 0 */
  most_common_group: number;
  most_common_mode: string;
  /** @minimum 0 */
  most_common_owner: number;
  owner_name?: string;
  path: string;
};